Move node dispatcher to class
