+++ src/configuration/registrations.ts
@@ -22,6 +22,7 @@ import { UpdateHandler } from '@/activity-handlers/update.handler';
 import { FedifyContextFactory } from '@/activitypub/fedify-context.factory';
 import { FediverseBridge } from '@/activitypub/fediverse-bridge';
 import { FollowersService } from '@/activitypub/followers.service';
+import { NodeInfoDispatcher } from '@/activitypub/nodeinfo.dispatcher';
 import { NodeInfoService } from '@/activitypub/nodeinfo.service';
 import { DeleteDispatcher } from '@/activitypub/object-dispatchers/delete.dispatcher';
 import type { ContextData } from '@/app';
@@ -39,7 +40,6 @@ import {
     createOutboxDispatcher,
     createUndoHandler,
     keypairDispatcher,
-    nodeInfoDispatcher,
 } from '@/dispatchers';
 import { EventSerializer } from '@/events/event';
 import { PubSubEvents } from '@/events/pubsub';
@@ -398,7 +398,7 @@ export function registerDependencies(
 
     container.register(
         'nodeInfoDispatcher',
-        asFunction(nodeInfoDispatcher).singleton(),
+        asClass(NodeInfoDispatcher).singleton(),
     );
 
     container.register(
