+++ src/dispatchers.ts
@@ -17,20 +17,14 @@ import {
     Like,
     Note,
     Person,
-    type Protocol,
     Undo,
     Update,
     verifyObject,
 } from '@fedify/fedify';
 import * as Sentry from '@sentry/node';
 
-import type { Account } from '@/account/account.entity';
 import type { AccountService } from '@/account/account.service';
 import type { FollowersService } from '@/activitypub/followers.service';
-import type {
-    NodeInfoData,
-    NodeInfoService,
-} from '@/activitypub/nodeinfo.service';
 import type { FedifyContext, FedifyRequestContext } from '@/app';
 import { ACTIVITYPUB_COLLECTION_PAGE_SIZE } from '@/constants';
 import { exhaustiveCheck, getError, getValue, isError } from '@/core/result';
@@ -1338,73 +1332,4 @@ export async function undoDispatcher(
     return Undo.fromJsonLd(exists);
 }
 
-export const nodeInfoDispatcher =
-    (
-        hostDataContextLoader: HostDataContextLoader,
-        nodeInfoService: NodeInfoService,
-    ) =>
-    async (ctx: FedifyRequestContext) => {
-        const hostData = await hostDataContextLoader.loadDataForHost(ctx.host);
 
-        if (isError(hostData)) {
-            ctx.data.logger.error('NodeInfo: failed to resolve host', {
-                host: ctx.host,
-                error: getError(hostData),
-            });
-            throw new Error('NodeInfo requested without site context');
-        }
-
-        const { site, account } = getValue(hostData);
-        const data = await nodeInfoService.getData(site, account);
-
-        return {
-            software: {
-                name: 'ghost' as const,
-                version: { major: 0, minor: 1, patch: 0 },
-                homepage: new URL('https://ghost.org/'),
-                repository: new URL('https://github.com/TryGhost/Ghost'),
-            },
-            protocols: ['activitypub'] as Protocol[],
-            services: {
-                inbound: [],
-                outbound: [],
-            },
-            openRegistrations: false,
-            usage: {
-                users: {
-                    total: 1,
-                    activeMonth: isActiveWithin(data.lastActivityAt, 30),
-                    activeHalfyear: isActiveWithin(data.lastActivityAt, 180),
-                },
-                localPosts: data.localPosts,
-                localComments: data.localComments,
-            },
-            metadata: getNodeInfoMetadata(account),
-        };
-    };
-
-function isActiveWithin(
-    lastActivityAt: NodeInfoData['lastActivityAt'],
-    days: number,
-): 0 | 1 {
-    if (lastActivityAt === null) {
-        return 0;
-    }
-
-    const activeSince = Date.now() - days * 24 * 60 * 60 * 1000;
-
-    return lastActivityAt.getTime() >= activeSince ? 1 : 0;
-}
-
-function getNodeInfoMetadata(account: Account) {
-    return {
-        nodeName: account.name ?? account.url.hostname,
-        ...(account.bio ? { nodeDescription: account.bio } : {}),
-        ...(account.avatarUrl ? { nodeIcon: account.avatarUrl.href } : {}),
-        ...(account.bannerImageUrl
-            ? { nodeBanner: account.bannerImageUrl.href }
-            : {}),
-        private: false,
-        postFormats: ['text/html'],
-    };
-}
